Update ConfigManager to produce destination and source mappings.
This is step 3 of the interval design. The configuration is now
parsed into unique destinations and sources. Historical Config
objects are treated as sources and the destination information
is pulled out of them and tracked. In this way the main thread
will know how many destination/source threads (and of what type) to
start.